+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 ## [Unreleased]
+### Added
+- Add optional deduplicate flag for Arweave API, default is on. Turn off to improve API response time. Estimate and Upload must be called with same deduplicate flag for price to work correctly
 
 ## [0.0.1]
 ### Added